chore: move actions to src
