+++ src/actions/verifyEmail.js
