+++ src/app/(auth)/verifyemail/page.jsx
@@ -1,4 +1,4 @@
-import { verifyEmail } from "@/app/actions/verifyEmail";
+import { verifyEmail } from "@/actions/verifyEmail";
 import Link from "next/link";
 
 const EmailVerificationPage = async (props) => {
